Add ISOMIP+ x/y to culled mesh
This requires a little reorganization of the function
`add_isomip_plus_xy()` that provides these coordinates on either
the planar or the spherical mesh.
