+++ polaris/ocean/tasks/isomip_plus/mesh/spherical.py
@@ -8,6 +8,7 @@ from mpas_tools.mesh.creation.signed_distance import (
 )
 
 from polaris.mesh import QuasiUniformSphericalMeshStep
+from polaris.ocean.tasks.isomip_plus.mesh.xy import add_isomip_plus_xy
 from polaris.ocean.tasks.isomip_plus.projection import get_projections
 
 
@@ -81,36 +82,6 @@ class SphericalMesh(QuasiUniformSphericalMeshStep):
         ds = xr.open_dataset('base_mesh_without_xy.nc')
         add_isomip_plus_xy(ds)
         ds.to_netcdf('base_mesh.nc')
-
-
-def add_isomip_plus_xy(ds):
-    """
-    Add x and y coordinates from a stereographic projection to a mesh on a
-    sphere
-
-    Parameters
-    ----------
-    ds : xarray.Dataset
-        The MPAS mesh on a sphere
-    """
-    projection, lat_lon_projection = get_projections()
-    transformer = pyproj.Transformer.from_proj(lat_lon_projection,
-                                               projection)
-    lon = np.rad2deg(ds.lonCell.values)
-    lat = np.rad2deg(ds.latCell.values)
-
-    x, y = transformer.transform(lon, lat)
-
-    ds['xIsomipCell'] = ('nCells', x)
-    ds['yIsomipCell'] = ('nCells', y)
-
-    lon = np.rad2deg(ds.lonVertex.values)
-    lat = np.rad2deg(ds.latVertex.values)
-
-    x, y = transformer.transform(lon, lat)
-
-    ds['xIsomipVertex'] = ('nVertices', x)
-    ds['yIsomipVertex'] = ('nVertices', y)
 
 
 def _make_feature(lx, ly, buffer):
